models/image/trufor/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic<2.0.0
# torch and torchvision are installed separately in Dockerfile for CPU
numpy==1.21.5 # Specific version from TruFor's test_docker/Dockerfile or conda.yaml
Pillow==10.0.1
python-multipart
timm==0.5.4 # Required by TruFor's SegFormer backbone
yacs>=0.1.8 # For config management used by TruFor
# gdown is not needed if wget is used for weights
opencv-python-headless # For OpenCV functionalities
scipy
scikit-learn
mmsegmentation==0.25.0
PyYAML==6.0